requirements.txt
flask==3.0.3
ollama==0.4.7
python-docx==1.1.2
python-dotenv==1.1.0
